refactor: start to implement MVC model
